requirements.txt
Django==2.2.6
django_ckeditor==5.6.1
Unidecode==1.0.22
django_debug_toolbar==1.9.1
django-dbbackup==3.2.0
dropbox==9.4.0
django-storages==1.7.2